server/requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
google-genai==1.5.0
requests==2.31.0
Pillow==11.1.0
